# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger API Documentation",
        "description": "Laravel",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registrar usuarios",
                "operationId": "App\\Http\\Controllers\\UserController::register",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "role_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario Registrado"
                    },
                    "422": {
                        "description": "Error de Validacion."
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "operationId": "App\\Http\\Controllers\\UserController::login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "422": {
                        "description": "Error de Validacion"
                    },
                    "200": {
                        "description": "Login Exitoso."
                    },
                    "401": {
                        "description": "Error de Credenciales."
                    }
                }
            }
        },
        "/api/refreshToken": {
            "post": {
                "tags": [
                    "Token"
                ],
                "summary": "Refresh Token",
                "operationId": "App\\Http\\Controllers\\UserController::refreshToken",
                "responses": {
                    "422": {
                        "description": "Error al refrescar el token"
                    },
                    "200": {
                        "description": "Refresh Token Exitoso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "operationId": "App\\Http\\Controllers\\UserController::logout",
                "responses": {
                    "422": {
                        "description": "Error al cerrar sesion"
                    },
                    "200": {
                        "description": "Logout Exitoso."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/getUserRole": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Obtener Rol del Usuario",
                "operationId": "App\\Http\\Controllers\\UserController::getUserRole",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exito"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/getUsers": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Obtener los Usuarios",
                "operationId": "App\\Http\\Controllers\\UserController::getUsers",
                "responses": {
                    "200": {
                        "description": "Exito"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}